fix: 통계 로직 최종 수정
