+++ src/main/java/com/barbet/howtodobe/domain/statistic/api/StatisticApi.java
