+++ src/main/java/com/barbet/howtodobe/domain/statistic/application/StatisticService.java
@@ -1,18 +1,15 @@
 package com.barbet.howtodobe.domain.statistic.application;
 
-import com.barbet.howtodobe.domain.category.dao.CategoryRepository;
 import com.barbet.howtodobe.domain.category.domain.Category;
 import com.barbet.howtodobe.domain.member.dao.MemberRepository;
 import com.barbet.howtodobe.domain.member.domain.Member;
-import com.barbet.howtodobe.domain.nowCategory.dao.NowCategoryRepository;
-import com.barbet.howtodobe.domain.nowCategory.domain.NowCategory;
-import com.barbet.howtodobe.domain.nowFailtag.dao.NowFailtagRepository;
-import com.barbet.howtodobe.domain.nowFailtag.domain.NowFailtag;
+import com.barbet.howtodobe.domain.statistic.domain.now.NowCategory;
+import com.barbet.howtodobe.domain.statistic.domain.now.NowFailtag;
 import com.barbet.howtodobe.domain.statistic.dto.StatisticResponseDTO;
 import com.barbet.howtodobe.domain.todo.dao.TodoRepository;
 import com.barbet.howtodobe.domain.todo.domain.Todo;
-import com.barbet.howtodobe.global.eunse.JwtTokenProvider;
-import com.barbet.howtodobe.global.exception.CustomException;
+import com.barbet.howtodobe.global.util.JwtTokenProvider;
+import com.barbet.howtodobe.global.common.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -23,7 +20,7 @@ import java.time.temporal.WeekFields;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static com.barbet.howtodobe.global.exception.CustomErrorCode.USER_NOT_FOUND;
+import static com.barbet.howtodobe.global.common.exception.CustomResponseCode.USER_NOT_FOUND;
 
 @Service
 @RequiredArgsConstructor
@@ -32,14 +29,11 @@ public class StatisticService {
     private final MemberRepository memberRepository;
 
     private final TodoRepository todoRepository;
-    private final CategoryRepository categoryRepository;
-    private final NowCategoryRepository nowCategoryRepository;
-    private final NowFailtagRepository nowFailtagRepository;
 
     /** 대분류 통계 정보 */
     private List<NowCategory> getWeekCategory(List<Todo> todoList) {
         Map<Category, Long> weekCategoryList = todoList.stream()
-                .filter(todo -> todo.getCategory().getCategoryId() != null)
+                .filter(todo -> todo.getCategory().getCategoryId() != null && todo.getIsChecked())
                 .collect(Collectors.groupingBy(Todo::getCategory, Collectors.counting()));
 
         Integer totalTodoCategoryCnt = todoList.size();
@@ -56,10 +50,6 @@ public class StatisticService {
                             .build();
                 })
                 .collect(Collectors.toList());
-
-        if (_weekCategory.size() > 3){
-            _weekCategory = _weekCategory.subList(0, 3);
-        }
         return _weekCategory;
     }
 
@@ -77,7 +67,6 @@ public class StatisticService {
                     String failtagName = entry.getKey();
                     Integer failtagCount = entry.getValue().intValue();
 
-                    // 실패 태그 비율을 전체 합에 대한 백분율로 계산
                     Integer nowFailtagRate = (totalTodoWithFailTagCnt != 0)
                             ? (int) ((failtagCount * 100.0) / totalFailtagRates)
                             : 0;
@@ -88,9 +77,6 @@ public class StatisticService {
                             .build();
                 })
                 .collect(Collectors.toList());
-        if (_weekFailtagList.size() > 3){
-            _weekFailtagList = _weekFailtagList.subList(0, 5);
-        }
         return _weekFailtagList;
     }
 
@@ -98,8 +84,7 @@ public class StatisticService {
     public StatisticResponseDTO getStatistic (LocalDate selectedDate, HttpServletRequest request) {
         Member member = memberRepository.findByMemberId(jwtTokenProvider.getUserId())
                 .orElseThrow(() -> new CustomException(USER_NOT_FOUND));
-        
-        // 회원인지 체크
+
         Long memberId = jwtTokenProvider.getUserIdByServlet(request);
         if (memberId != null && !memberId.equals(member.getMemberId())) {
             throw new CustomException(USER_NOT_FOUND);
@@ -154,7 +139,6 @@ public class StatisticService {
                     .collect(Collectors.toList());
             nowWorstFailTag = nowFailTagData.get(0).getNowFailtag();
         }
-
         return new StatisticResponseDTO(selectedDate, prevTodoCnt, prevTodoDoneCnt, nowTodoCnt, nowTodoDoneCnt, rateOfChange, nowCategoryData, nowBestCateogry, nowFailTagData, nowWorstFailTag);
     }
 }
